refactor(keystore-card): move component and story to general
